requirements.txt
flask
gunicorn
pandas
numpy
scikit-learn
mysql-connector-python